float cast
git-svn-id: https://jumble.svn.sourceforge.net/svnroot/jumble/trunk@442 3c96dea9-a453-4bd5-b869-110af7cb2dd7

+++ src/jumble/mutation/Mutater.java
@@ -739,7 +739,7 @@ public class Mutater {
     } else if (c instanceof ConstantFloat) {
       final float current = ((ConstantFloat) c).getBytes();
       // can't just add 1 because of potential underflow
-      final double newValue = 2 * current + 1;
+      final float newValue = (float) (2 * current + 1);
       cp.setConstant(i, new ConstantFloat(newValue));
       mod = mod + current + " -> " + newValue;
     } else if (c instanceof ConstantDouble) {
